refactor: Extract functions from generate.rs to utils.rs
